pyproject.toml
[project]
name = "acta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=6.0.0",
    "matplotlib>=3.10.7",
    "mesa>=3.3.1",
    "networkx>=3.6",
    "pyyaml>=6.0.3",
    "solara>=1.55.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true
